+++ addons/yat/src/commands/Ping.cs
@@ -14,25 +14,25 @@ namespace YAT.Commands;
 )]
 [Threaded]
 [Argument("hostname", "string", "The host to trace the route to.")]
-[Option("-t", "int(1:32767)", "The maximum time to wait for each reply, in milliseconds.", 10000)]
-[Option("-ttl", "int(1:255)", "The maximum number of hops to search for the target.", 30)]
-[Option("-b", "int(1:32767)", "The size of the buffer to send with the request.", 32)]
+[Option("-t", "int(1:32767)", "The maximum time to wait for each reply, in milliseconds.", 10000f)]
+[Option("-ttl", "int(1:255)", "The maximum number of hops to search for the target.", 30f)]
+[Option("-b", "int(1:32767)", "The size of the buffer to send with the request.", 32f)]
 [Option("-f", "bool", "Specifies that the packet can be fragmented.", false)]
-[Option("-delay", "int(1:10)", "The delay between pings in seconds.", 1)]
-[Option("-limit", "int(1:255)", "The maximum number of pings to send.", 0)]
+[Option("-delay", "int(1:10)", "The delay between pings in seconds.", 1f)]
+[Option("-limit", "int(1:255)", "The maximum number of pings to send.", 0f)]
 public sealed class Ping : ICommand
 {
 	public CommandResult Execute(CommandData data)
 	{
 		string hostname = (string)data.Arguments["hostname"];
-		var maxPings = (int)data.Options["-limit"];
+		var maxPings = (int)(float)data.Options["-limit"];
 		var options = new NetworkingOptions
 		{
-			Timeout = (ushort)(int)data.Options["-t"],
-			TTL = (ushort)(int)data.Options["-ttl"],
-			BufferSize = (ushort)(int)data.Options["-b"],
+			Timeout = (ushort)(float)data.Options["-t"],
+			TTL = (ushort)(float)data.Options["-ttl"],
+			BufferSize = (ushort)(float)data.Options["-b"],
 			DontFragment = !(bool)data.Options["-f"],
-			Delay = (ushort)((int)data.Options["-delay"] * 1000),
+			Delay = (ushort)((float)data.Options["-delay"] * 1000),
 		};
 
 		uint pings = 0;
