+++ addons/yat/src/commands/TraceRoute.cs
@@ -12,9 +12,9 @@ namespace YAT.Commands;
 	"[b]Usage[/b]: traceroute [i]hostname[/i]"
 )]
 [Argument("hostname", "string", "The host to trace the route to.")]
-[Option("-t", "int(1:32767)", "The maximum time to wait for each reply, in milliseconds.", 10000)]
-[Option("-ttl", "int(1:255)", "The maximum number of hops to search for the target.", 30)]
-[Option("-b", "int(1:32767)", "The size of the buffer to send with the request.", 32)]
+[Option("-t", "int(1:32767)", "The maximum time to wait for each reply, in milliseconds.", 10000f)]
+[Option("-ttl", "int(1:255)", "The maximum number of hops to search for the target.", 30f)]
+[Option("-b", "int(1:32767)", "The size of the buffer to send with the request.", 32f)]
 [Option("-f", "bool", "Specifies that the packet can be fragmented.", false)]
 [Threaded]
 public sealed class TraceRoute : ICommand
@@ -24,9 +24,9 @@ public sealed class TraceRoute : ICommand
 		var hostname = (string)data.Arguments["hostname"];
 		var options = new NetworkingOptions
 		{
-			Timeout = (ushort)(int)data.Options["-t"],
-			TTL = (ushort)(int)data.Options["-ttl"],
-			BufferSize = (ushort)(int)data.Options["-b"],
+			Timeout = (ushort)(float)data.Options["-t"],
+			TTL = (ushort)(float)data.Options["-ttl"],
+			BufferSize = (ushort)(float)data.Options["-b"],
 			DontFragment = !(bool)data.Options["-f"]
 		};
